requirements.txt
streamlit
pandas
numpy
matplotlib
wordcloud
Pillow
